refactor: 💡 loading text change
